refactored Flashcard View
