+++ frontend/frontend_2/src/App.js
@@ -6,7 +6,7 @@ import useToken from './useToken';
 import Login from './components/views/auth/Login';
 import Register from './components/views/auth/Register';
 import Logout from './components/views/auth/Logout'
-import FlashcardView from './components/views/FlashCardView'
+import FlashcardView from './components/views/FlashcardView/FlashCardView'
 import FlashcardOptions from './components/views/FlashcardView/FlashcardOptions.js'
 import 'bootstrap/dist/css/bootstrap.css';
