+++ frontend/frontend_2/src/components/views/FlashcardView/FlashCardView.js
